Moved platform tests to utils file
